chore(api): moving ban and revocation endpoints to internal API
