+++ api/ceramic_cache/api/v1.py
@@ -50,11 +50,10 @@ from trusta_labs.api import CgrantsApiKey
 
 from ..exceptions import (
     InternalServerException,
-    InvalidBanQueryException,
     InvalidDeleteCacheRequestException,
     TooManyStampsException,
 )
-from ..models import Ban, CeramicCache, Revocation
+from ..models import CeramicCache
 from ..utils import validate_dag_jws_payload
 from .schema import (
     AccessTokenResponse,
@@ -62,15 +61,11 @@ from .schema import (
     CachedStampResponse,
     CacheStampPayload,
     CalcScorePayload,
-    CheckBanResult,
     ComporeDBStatusResponse,
     ComposeDBStatusPayload,
-    Credential,
     DeleteStampPayload,
     GetStampResponse,
     GetStampsWithScoreResponse,
-    RevocationCheckPayload,
-    RevocationCheckResponse,
 )
 
 log = logging.getLogger(__name__)
@@ -710,89 +705,3 @@ def accept_tos(
     request, tos_type: str, address: str, payload: tos.schema.TosSigned
 ) -> None:
     tos.api.accept_tos(payload)
-
-
-@router.post("/check-bans", response=List[CheckBanResult], auth=secret_key)
-def check_bans(request, payload: List[Credential]) -> List[CheckBanResult]:
-    """
-    Check for active bans matching the given address and/or hashes.
-    Returns list of relevant active bans.
-    """
-    unique_addresses = set([c.credentialSubject.address for c in payload])
-
-    if len(unique_addresses) < 1:
-        raise InvalidBanQueryException("Must provide valid credential(s)")
-
-    if len(unique_addresses) > 1:
-        raise InvalidBanQueryException(
-            "All credentials must be issued to the same address"
-        )
-
-    address = list(unique_addresses)[0]
-
-    hashes = list(
-        set([c.credentialSubject.hash for c in payload if c.credentialSubject.hash])
-    )
-
-    try:
-        bans = Ban.get_bans(address=address, hashes=hashes)
-
-        credential_ban_results = [
-            Ban.check_bans_for(
-                bans, address, c.credentialSubject.hash, c.credentialSubject.provider
-            )
-            for c in payload
-        ]
-
-        return [
-            CheckBanResult(
-                credential_id=c.id,
-                is_banned=is_banned,
-                ban_type=ban_type,
-                end_time=ban.end_time if ban else None,
-                reason=ban.reason if ban else None,
-            )
-            for c, (is_banned, ban_type, ban) in zip(payload, credential_ban_results)
-        ]
-
-    except APIException:
-        # re-raise API exceptions
-        raise
-    except Exception as e:
-        log.error("Failed to check bans", exc_info=True)
-        raise InternalServerException("Failed to check bans") from e
-
-
-@router.post("/check-revocations", response=List[RevocationCheckResponse], auth=None)
-def check_revocations(
-    request, payload: RevocationCheckPayload
-) -> List[RevocationCheckResponse]:
-    """
-    Check if stamps with given proof values have been revoked.
-    Returns revocation status for each proof value.
-    """
-    if len(payload.proof_values) > settings.MAX_BULK_CACHE_SIZE:
-        raise TooManyStampsException()
-
-    try:
-        # Query for revocations matching any of the proof values
-        revoked_proof_values = set(
-            Revocation.objects.filter(proof_value__in=payload.proof_values).values_list(
-                "proof_value", flat=True
-            )
-        )
-
-        # Return status for each requested proof value
-        return [
-            RevocationCheckResponse(
-                proof_value=proof_value, is_revoked=proof_value in revoked_proof_values
-            )
-            for proof_value in payload.proof_values
-        ]
-
-    except APIException:
-        # re-raise API exceptions
-        raise
-    except Exception as e:
-        log.error("Failed to check revocations", exc_info=True)
-        raise InternalServerException("Failed to check revocations") from e
